test(lsp): isolate launcher fingerprint parity
